backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Session==0.5.0
python-dotenv==1.0.0
PyJWT==2.8.1
requests==2.31.0
google-auth==2.26.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.107.0
